+++ src/Servant/ActiveResource.hs
@@ -5,26 +5,24 @@
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 
-
 -- |
 --
 -- Module      : Servant.ActiveResource
--- Copyright   : (C) 2024 Bellroy Pty Ltd
+-- Copyright   : (C) 2024-2025 Bellroy Pty Ltd
 -- License     : BSD-3-Clause
 -- Maintainer  : Bellroy Tech Team <haskell@bellroy.com>
 -- Stability   : experimental
 --
 -- Types and helpers for defining Servant routes compatible with
--- [Rails' ActiveResource](https://github.com/rails/activeresource).
+-- [Rails's ActiveResource](https://github.com/rails/activeresource).
 --
 -- @
--- {-# LANGUAGE TemplateHaskell #-}
---
 -- import qualified Servant.ActiveResource as AR
 --
 -- newtype MyResourceId = MyResourceId Int
@@ -34,31 +32,49 @@
 -- -- Like MyResource, but returned from the database.
 -- data MyStoredResource = MyStoredResource {...}
 --
--- -- The exact monad used will depend on your program. Here, we just assume
--- -- 'Handler' from package servant-server.
--- instance AR.'Resource' MyResourceId Handler where
---   type 'ResourceData' MyResourceId = MyResource
---   type 'StoredResourceData' MyResourceId = MyStoredResource
+-- -- These type family instances associate your resource's ID type
+-- -- with the data types accepted and returned by your operations and
+-- -- by the servant-server API.
+-- type instance 'ResourceData' MyResourceId = MyResource
+-- type instance 'StoredResourceData' MyResourceId = MyStoredResource
 --
---   -- These form the implementation of your API.
---   'listResources' = ...
---   'createResource' = ...
---   'readResource' = ...
---   'upsertResource' = ...
---   'deleteResource' = ...
---
--- -- Record of routes, which can be spliced into a top-level handler
--- -- via 'Servant.API.NamedRoutes'.
--- routes :: AR.'ResourceRoutes' MyResourceId (AsServerT Handler)
--- routes = $(AR.'makeResourceServerT' [t|MyResourceId|])
+-- -- Record of routes, which can passed directly to
+-- -- 'Servant.Server.Generic.genericServe', or spliced into another
+-- -- record of routes via 'Servant.API.NamedRoutes'.
+-- --
+-- -- The exact monad used will depend on your program. Here, we just
+-- -- assume 'Handler' from package servant-server.
+-- routes :: AR.'ResourceRoutes' MyResourceId AsServer
+-- routes = AR.'makeResourceRoutes' 'ResourceOperations'
+--   { 'listResources' = ...
+--   , 'createResource' = ...
+--   , 'readResource' = ...
+--   , 'upsertResource' = ...
+--   , 'deleteResource' = ...
+--   }
 -- @
 module Servant.ActiveResource
-  ( Resource (..),
+  ( -- * Declaring Resources
+    ResourceData,
+    StoredResourceData,
+    ResourceOperations (..),
+
+    -- ** Types used in CRUDL operation results
     NotFoundError (..),
     ValidationError (..),
     CreatedUpdated (..),
+
+    -- ** Helpers
+    hoistResourceOperations,
+
+    -- * Types for @servant@ and @servant-server@
     ResourceRoutes (..),
-    makeResourceServerT,
+
+    -- ** Creating 'ResourceRoutes'
+    makeResourceRoutes,
+    makeResourceRoutesT,
+
+    -- * Miscellaneous
     errorFormatters,
   )
 where
@@ -72,58 +88,71 @@ import Data.Map (Map)
 import Data.Proxy (Proxy (..))
 import Data.Text (Text)
 import GHC.Generics (Generic)
-import qualified Language.Haskell.TH as TH
 import Servant.API
   ( Capture,
     JSON,
-    NoContent (..), -- ctor import needed by TH
+    NoContent (..),
     ReqBody,
     StdMethod (..),
     UVerb,
-    WithStatus (..), -- ctor import needed by TH
+    WithStatus (..),
     (:>),
   )
 import Servant.API.ContentTypes (handleAcceptH)
 import Servant.API.Generic ((:-))
 import Servant.Server
   ( ErrorFormatters (..),
+    Handler,
     defaultErrorFormatters,
     err422,
     errBody,
     errHeaders,
     getAcceptHeader,
-    respond, -- needed by TH
+    respond,
   )
+import Servant.Server.Generic (AsServer, AsServerT)
 
--- | A 'Resource' instance says that the type @id@ is the primary
--- key for some abstract resource, and that the collection of such
--- resources can be manipulated with CRUDL actions inside a Monad @m@.
-class (Monad m) => Resource id m where
-  -- | The type used to represent new records or updates to existing
-  -- records. Typically, this is missing an @id@ field.
-  type ResourceData id :: Type
+-- | The type used to represent new records or updates to existing
+-- records. Typically, this is missing an @id@ field.
+--
+-- @since 0.2.0.0
+type family ResourceData id :: Type
 
-  -- | The type used to represent existing records.
-  type StoredResourceData id :: Type
+-- | The type used to represent existing records.
+--
+-- @since 0.2.0.0
+type family StoredResourceData id :: Type
 
-  listResources ::
-    m [StoredResourceData id]
-  createResource ::
-    ResourceData id ->
-    m (Either ValidationError (StoredResourceData id))
-  readResource ::
-    id ->
-    m (Either NotFoundError (StoredResourceData id))
-  upsertResource ::
-    id ->
-    ResourceData id ->
-    m (Either ValidationError (CreatedUpdated, StoredResourceData id))
-  deleteResource ::
-    id ->
-    m (Either NotFoundError ())
+-- | A 'ResourceOperations id m' record collects the implementations of CRUDL
+-- handlers for a resource.
+--
+-- The type @id@ is the primary key for some abstract resource, and
+-- operations on the collection of such resources all occur inside a
+-- Monad @m@.
+--
+-- @since 0.2.0.0
+data ResourceOperations id m = ResourceOperations
+  { listResources ::
+      m [StoredResourceData id],
+    createResource ::
+      ResourceData id ->
+      m (Either ValidationError (StoredResourceData id)),
+    readResource ::
+      id ->
+      m (Either NotFoundError (StoredResourceData id)),
+    upsertResource ::
+      id ->
+      ResourceData id ->
+      m (Either ValidationError (CreatedUpdated, StoredResourceData id)),
+    deleteResource ::
+      id ->
+      m (Either NotFoundError ())
+  }
+  deriving stock (Generic)
 
+-- | @since 0.1.0.0
 data NotFoundError = NotFoundError
-  deriving (Eq, Ord, Bounded, Show, Generic)
+  deriving stock (Eq, Ord, Bounded, Show, Generic)
 
 instance ToJSON NotFoundError where
   toJSON NotFoundError =
@@ -136,23 +165,44 @@ instance ToJSON NotFoundError where
 --
 -- The 'ToJSON' instance wraps the whole thing in an @{ "errors": ... }@
 -- object, to match [ActiveResource's expectations](https://github.com/rails/activeresource/blob/3cf44f731f655dccc13ba23f78603f7e214e3352/lib/active_resource/validations.rb#L35-L71).
+--
+-- @since 0.1.0.0
 newtype ValidationError = ValidationError (Map Text [Text])
   deriving stock (Eq, Show, Generic)
 
 instance ToJSON ValidationError where
   toJSON (ValidationError errs) = object ["errors" .= toJSON errs]
 
+-- | @since 0.1.0.0
 data CreatedUpdated = Created | Updated
-  deriving (Eq, Ord, Bounded, Show, Generic)
+  deriving stock (Eq, Ord, Bounded, Show, Generic)
+
+-- | Apply a natural tranformation to the CRUDL endpoints for a
+-- resource.
+--
+-- @since 0.2.0.0
+hoistResourceOperations ::
+  (forall x. m x -> n x) -> ResourceOperations id m -> ResourceOperations id n
+hoistResourceOperations nt ResourceOperations {..} =
+  ResourceOperations
+    { listResources = nt listResources,
+      createResource = nt . createResource,
+      readResource = nt . readResource,
+      upsertResource = (nt .) . upsertResource,
+      deleteResource = nt . deleteResource
+    }
 
 -- | The CRUDL routes required by an ActiveResource API, using Servant's
 -- <https://docs.servant.dev/en/stable/cookbook/generic/Generic.html generic record-based routing>.
--- Once you have a 'Resource' instance, use 'makeResourceServerT' to
--- generate a server that fills out this structure.
+-- The easiest way to create one of these is to implement the CRUDL
+-- operations in a 'ResourceOperations' record, and then apply
+-- 'makeResourceRoutesT' or 'makeResourceRoutes' to it.
 --
 -- NOTE: If the
 -- <https://github.com/haskell-servant/servant/issues/1466 alternate syntax>
 -- for ':-' gets merged, we intend to adopt it; it's easier to read.
+--
+-- @since 0.1.0.0
 data ResourceRoutes (id :: Type) mode = ResourceRoutes
   { listRoute ::
       mode
@@ -202,55 +252,53 @@ data ResourceRoutes (id :: Type) mode = ResourceRoutes
   }
   deriving (Generic)
 
--- | Given an instance of 'ResourceId ty', plumb through its operations to
--- build a 'ResourceRoutes'. GHC's type checker
--- <https://github.com/haskell-servant/servant/issues/1381 isn't smart enough>
--- to reason through the generated code if it's given a polymorphic
--- type, so we have to use TH.
+-- | Turn a record of CRUDL operations operating in @servant-server@'s
+-- 'Handler' monad into a record of endpoint handlers, suitable for
+-- use with @servant-server@'s 'Servant.Server.Generic.NamedRoutes'.
 --
--- If @m ~ ResourceMonad ty@, then you can call it like this:
+-- Most of the time, you'll use 'makeResourceRoutesT', which works in
+-- your application's monad.
 --
--- @
--- myResourceServer :: 'ResourceRoutes' ty ('Servant.Generic.AsServerT' m)
--- myResourceServer = \$(makeResourceServerT [t|ty|])
--- @
+-- @since 0.2.0.0
+makeResourceRoutes ::
+  ResourceOperations id Handler -> ResourceRoutes id AsServer
+makeResourceRoutes = makeResourceRoutesT
+
+-- | Turn a record of CRUDL operations into a actual record of
+-- endpoint handlers. This works for any application monad.
 --
--- The generated code will
--- need @-XDataKinds@, @-XLambdaCase@, and @-XTypeApplications@.
---
--- You can then serve it using the functions in "Servant.Server.Generic".
-makeResourceServerT :: TH.TypeQ -> TH.ExpQ
-makeResourceServerT ty =
-  [|
-    ResourceRoutes
-      { listRoute =
-          $(TH.appTypeE [|listResources|] ty)
-            >>= $(respondWithStatus 200),
-        createRoute =
-          $(TH.appTypeE [|createResource|] ty)
-            >=> either $(respondWithStatus 422) $(respondWithStatus 201),
-        readRoute =
-          $(TH.appTypeE [|readResource|] ty)
-            >=> either $(respondWithStatus 404) $(respondWithStatus 200),
-        upsertRoute = \id_ data_ ->
-          $(TH.appTypeE [|upsertResource|] ty) id_ data_
-            >>= either
-              $(respondWithStatus 422)
-              ( \case
-                  (Created, r) -> $(respondWithStatus 201) r
-                  (Updated, r) -> $(respondWithStatus 200) r
-              ),
-        deleteRoute =
-          $(TH.appTypeE [|deleteResource|] ty)
-            >=> either
-              $(respondWithStatus 404)
-              (const ($(respondWithStatus 204) NoContent))
-      }
-    |]
-  where
-    -- respondWithStatus n ===> respond . WithStatus @n
-    respondWithStatus :: Integer -> TH.ExpQ
-    respondWithStatus n = [|respond . $(TH.appTypeE [|WithStatus|] (TH.litT (TH.numTyLit n)))|]
+-- @since 0.2.0.0
+makeResourceRoutesT ::
+  (Monad m) => ResourceOperations id m -> ResourceRoutes id (AsServerT m)
+makeResourceRoutesT ResourceOperations {..} =
+  ResourceRoutes
+    { listRoute =
+        listResources
+          >>= respond . WithStatus @200,
+      createRoute =
+        createResource
+          >=> either
+            (respond . WithStatus @422)
+            (respond . WithStatus @201),
+      readRoute =
+        readResource
+          >=> either
+            (respond . WithStatus @404)
+            (respond . WithStatus @200),
+      upsertRoute = \id_ data_ ->
+        upsertResource id_ data_
+          >>= either
+            (respond . WithStatus @422)
+            ( \case
+                (Created, r) -> respond $ WithStatus @201 r
+                (Updated, r) -> respond $ WithStatus @200 r
+            ),
+      deleteRoute =
+        deleteResource
+          >=> either
+            (respond . WithStatus @404)
+            (const (respond $ WithStatus @204 NoContent))
+    }
 
 -- | This is a custom error formatter for parse errors, which runs when
 -- Servant fails to deserialise a request parameter (i.e., before endpoint
@@ -269,6 +317,8 @@ makeResourceServerT ty =
 -- -- can pass 'id' as the natural transformation parameter:
 -- 'Servant.Server.Generic.genericServeTWithContext' id myResourceServer (errorFormatters ':.' 'EmptyContext')
 -- @
+--
+-- @since 0.1.0.0
 errorFormatters :: ErrorFormatters
 errorFormatters =
   defaultErrorFormatters